chatbot_service/requirements.txt


python-dotenv


psycopg2-binary


langchain-openai
openai



python-docx




fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
python-docx
pypdf2

langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
